Solution to Print a Pyramid Pattern with Numbers

+++ Print a Pyramid Pattern with Numbers@@67877e1818d53eb724f435c6/code.c
@@ -1,7 +1,7 @@
 // Your code here...
 #include <stdio.h>
 int factorial(int num){
-    int fac=1;
+    double fac=1;
     for(int i=1;i<=num;i++){
         fac*=i;
     }
